requirements.txt
edge_tts==6.1.12
PyAudio==0.2.14
pydub==0.25.1
assemblyai
ollama
google-cloud-texttospeech
azure-cognitiveservices-speech
SpeechRecognition